chore: modify Extract the type
