+++ package.json
@@ -2,10 +2,11 @@
   "name": "garfish",
   "version": "0.0.3",
   "private": true,
-  "workspaces": [
-    "packages/runtime/*",
-    "packages/tools/*"
-  ],
+  "workspaces": {
+    "packages": [
+      "packages/core/*"
+    ]
+  },
   "description": "a powerful micro front-end framework",
   "scripts": {
     "dev": "node scripts/dev.js",
@@ -26,12 +27,6 @@
     "format": "prettier --write --parser typescript \"packages/**/*.ts?(x)\"",
     "format:js": "prettier --write --parser typescript \"(packages|scripts)/**/*.js\"",
     "format:md": "prettier --write \"**/*.md\""
-  },
-  "private": true,
-  "workspaces": {
-    "packages": [
-      "packages/core/*"
-    ]
   },
   "husky": {
     "hooks": {
