+++ packages/runtime/browser-snapshot/api-extractor.json
@@ -1,8 +1,7 @@
 {
   "extends": "../../../api-extractor.json",
   "mainEntryPointFilePath": "./dist/packages/runtime/browser-snapshot/src/index.d.ts",
-  "globalEntryPointFilePath": "./dist/packages/runtime/browser-snapshot/src/global.d.ts",
   "dtsRollup": {
-    "publicTrimmedFilePath": "./dist/browser-snapshot.d.ts"
+    "publicTrimmedFilePath": "./dist/index.d.ts"
   }
 }
