+++ packages/runtime/browser-snapshot/src/globalExtensions.ts
@@ -1,0 +1,19 @@
+import { SnapshotSandbox } from './sandbox';
+
+declare module '@garfish/core' {
+  export namespace interfaces {
+    export interface Config {
+      protectVariable?: PropertyKey[];
+      insulationVariable?: PropertyKey[];
+      sandbox?: SandboxConfig;
+    }
+
+    export interface App {
+      snapshotSandbox?: SnapshotSandbox;
+    }
+
+    export interface Plugin {
+      openBrowser?: boolean;
+    }
+  }
+}
