+++ packages/runtime/browser-vm/api-extractor.json
@@ -2,6 +2,6 @@
   "extends": "../../../api-extractor.json",
   "mainEntryPointFilePath": "./dist/packages/runtime/browser-vm/src/index.d.ts",
   "dtsRollup": {
-    "publicTrimmedFilePath": "./dist/browser-vm.d.ts"
+    "publicTrimmedFilePath": "./dist/index.d.ts"
   }
 }
