+++ packages/runtime/core/package.json
@@ -34,6 +34,7 @@
   "dependencies": {
     "@garfish/hooks": "^0.0.9-alpha.13",
     "@garfish/utils": "^0.0.9-alpha.13",
+    "@garfish/loader": "^0.0.9-alpha.13",
     "himalaya": "^1.1.0"
   },
   "publishConfig": {
