+++ packages/runtime/garfish/api-extractor.json
@@ -2,6 +2,6 @@
   "extends": "../../../api-extractor.json",
   "mainEntryPointFilePath": "./dist/packages/runtime/garfish/src/index.d.ts",
   "dtsRollup": {
-    "publicTrimmedFilePath": "./dist/garfish.d.ts"
+    "publicTrimmedFilePath": "./dist/index.d.ts"
   }
 }
