+++ packages/runtime/garfish/index.d.ts
@@ -1,0 +1,6 @@
+/// <reference types="@garfish/browser-vm" />
+/// <reference types="@garfish/browser-snapshot" />
+/// <reference types="@garfish/router" />
+export * from './dist/index';
+import defaultInstance from './dist/index';
+export default defaultInstance;
