+++ packages/runtime/garfish/package.json
@@ -19,7 +19,7 @@
     "url": "https://github.com/bytedance/garfish/issues"
   },
   "scripts": {},
-  "types": "dist/index.d.ts",
+  "types": "./index.d.ts",
   "buildOptions": {
     "name": "Garfish",
     "devTemplate": "module",
