+++ packages/runtime/router/api-extractor.json
@@ -2,6 +2,6 @@
   "extends": "../../../api-extractor.json",
   "mainEntryPointFilePath": "./dist/packages/runtime/router/src/index.d.ts",
   "dtsRollup": {
-    "publicTrimmedFilePath": "./dist/router.d.ts"
+    "publicTrimmedFilePath": "./dist/index.d.ts"
   }
 }
