+++ packages/runtime/router/src/index.ts
@@ -2,37 +2,14 @@ import { interfaces } from '@garfish/core';
 import { createKey } from '@garfish/utils';
 import router, {
   initRedirect,
-  RouterInterface,
   listenRouterAndReDirect,
+  RouterInterface,
 } from './context';
+import './globalExtensions';
 
 interface Options {
   autoRefreshApp?: boolean;
   onNotMatchRouter?: (path: string) => Promise<void> | void;
-}
-
-declare module '@garfish/core' {
-  export interface Garfish {
-    router: RouterInterface;
-  }
-
-  export namespace interfaces {
-    export interface Garfish {
-      router: RouterInterface;
-    }
-
-    export interface Config {
-      autoRefreshApp?: boolean;
-      onNotMatchRouter?: (path: string) => Promise<void> | void;
-    }
-
-    export interface AppInfo {
-      activeWhen?: string | ((path: string) => boolean); // 手动加载，可不填写路由
-      active?: (appInfo: AppInfo, rootPath: string) => void;
-      deactive?: (appInfo: AppInfo, rootPath: string) => void;
-      basename?: string;
-    }
-  }
 }
 
 export default function Router(_args?: Options) {
@@ -121,3 +98,5 @@ export default function Router(_args?: Options) {
     };
   };
 }
+
+export { RouterInterface } from './context';
