+++ packages/runtime/utils/api-extractor.json
@@ -2,6 +2,6 @@
   "extends": "../../../api-extractor.json",
   "mainEntryPointFilePath": "./dist/packages/runtime/utils/src/index.d.ts",
   "dtsRollup": {
-    "publicTrimmedFilePath": "./dist/utils.d.ts"
+    "publicTrimmedFilePath": "./dist/index.d.ts"
   }
 }
